requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiofiles==24.1.0
    # via frontend
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
beautifulsoup4==4.13.4
    # via -r requirements.in
blinker==1.9.0
    # via flask
blis==1.3.0
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   flask
    #   typer
    #   uvicorn
cloudpathlib==0.21.0
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cryptography==45.0.4
    # via -r requirements.in
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.1.1
    # via -r requirements.in
frontend==0.0.3
    # via -r requirements.in
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
huggingface-hub==0.33.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via
    #   flask
    #   frontend
jinja2==3.1.6
    # via
    #   flask
    #   spacy
    #   torch
jiter==0.9.0
    # via openai
joblib==1.5.1
    # via scikit-learn
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
murmurhash==1.0.12
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==3.4.2
    # via torch
numpy==2.0.2
    # via
    #   blis
    #   pandas
    #   pgvector
    #   scikit-learn
    #   scipy
    #   spacy
    #   thinc
    #   transformers
openai==1.77.0
    # via -r requirements.in
packaging==25.0
    # via
    #   huggingface-hub
    #   spacy
    #   thinc
    #   transformers
    #   weasel
pandas==2.2.3
    # via -r requirements.in
pgvector==0.4.1
    # via -r requirements.in
pillow==11.2.1
    # via sentence-transformers
preshed==3.0.9
    # via
    #   spacy
    #   thinc
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.11.4
    # via
    #   confection
    #   fastapi
    #   openai
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via -r requirements.in
pymupdf==1.26.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   spacy
    #   transformers
    #   weasel
rich==14.0.0
    # via typer
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via sentence-transformers
scipy==1.15.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.1.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smart-open==7.1.0
    # via weasel
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
spacy==3.8.5
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
sqlalchemy==1.4.31
    # via -r requirements.in
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
starlette==0.46.2
    # via
    #   fastapi
    #   frontend
sympy==1.14.0
    # via torch
thinc==8.3.6
    # via spacy
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
tools==1.0.2
    # via -r requirements.in
torch==2.7.1
    # via
    #   -r requirements.in
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   spacy
    #   transformers
transformers==4.52.4
    # via sentence-transformers
typer==0.15.3
    # via
    #   spacy
    #   weasel
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   cloudpathlib
    #   fastapi
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich
    #   sentence-transformers
    #   torch
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via frontend
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
werkzeug==3.1.3
    # via flask
wrapt==1.17.2
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
